+++ src/cocotb/regression.py
@@ -126,8 +126,8 @@ class Test:
             Units of ``timeout_time``, accepts any units that :class:`~cocotb.triggers.Timer` does.
 
         expect_fail:
-            If ``True`` and the test fails a functional check via an ``assert`` statement, :pytest:class:`pytest.raises`,
-            :pytest:class:`pytest.warns`, or :pytest:class:`pytest.deprecated_call`, the test is considered to have passed.
+            If ``True`` and the test fails a functional check via an ``assert`` statement, :func:`pytest.raises`,
+            :func:`pytest.warns`, or :func:`pytest.deprecated_call`, the test is considered to have passed.
             If ``True`` and the test passes successfully, the test is considered to have failed.
 
         expect_error: